Add initial version of a position scrapping script
This is a script which tries to fetch all bus positions for all lines
(supposedly) in the SPTrans Olho Vivo API every second (or half a
second), piping the data into MongoDB.

I'm unsure about the database solution, as well as the script's error
handling. But we'll fix these problems with time. Ideally this script
wouldn't really be written in Node.JS as single-threading/processing
might be a problem on the long run.